+++ src/ui-stack-layout/ui-stack-layout-page.d.ts
@@ -1,5 +1,11 @@
 export class UIStackLayoutPage extends HTMLElement {
     static register: () => void;
+    static observedAttributes: string[];
+    /**
+     * @param {string} name
+     */
+    constructor(name: string);
+    name: string;
     ui: {
         root: this;
         name: string;
@@ -7,4 +13,10 @@ export class UIStackLayoutPage extends HTMLElement {
     shadowRender(): void;
     connectedCallback(): void;
     disconnectedCallback(): void;
+    /**
+     * @param {string} n
+     * @param {string | null} _oV
+     * @param {string | null} nV
+     */
+    attributeChangedCallback(n: string, _oV: string | null, nV: string | null): void;
 }